PET-115 test : TodoControllerTest 작성, TodoHomeControllerTest 삭제, API 명세 반영
