+++ Domain-Module/Todo-Module/Todo-Presentation/src/docs/asciidoc/Todo-API.adoc
@@ -28,16 +28,16 @@ operation::todo-team-controller-test/post-todo-team[snippets='http-request,reque
 operation::todo-team-controller-test/get-todo-team-random-code[snippets='http-request,request-headers,http-response,response-fields']
 
 [[Todo-API-Todo-달성률]]
-=== Todo 달성률 (홈 화면)
+=== Todo 달성률 
 
-operation::todo-home-controller-test/get-todo-progress[snippets='http-request,path-parameters,request-headers,http-response,response-fields']
+operation::todo-controller-test/get-todo-progress[snippets='http-request,path-parameters,request-headers,http-response,response-fields']
 
 [[Todo-API-가장-최근에-가입한-Todo-Team-조회]]
-=== 가장 최근에 가입한 TodoTeam 조회 (홈 화면)
+=== 가장 최근에 가입한 TodoTeam 조회
 
-operation::todo-home-controller-test/get-todo-team-name[snippets='http-request,request-headers,http-response,response-fields']
+operation::todo-team-controller-test/get-todo-team-name[snippets='http-request,request-headers,http-response,response-fields']
 
 [[Todo-API-할당-받은-Todo-조회]]
 === 할당받은 Todo 조회
 
-operation::todo-home-controller-test/get-todos[snippets='http-request,path-parameters,request-parameters,request-headers,http-response,response-fields']
+operation::todo-controller-test/get-todos[snippets='http-request,path-parameters,request-parameters,request-headers,http-response,response-fields']
